Split Symtab into slices
