--- rsc/src/main/scala/rsc/pretty/PrettySymtab.scala
@@ -1,15 +1,0 @@
-// Copyright (c) 2017-2019 Twitter, Inc.
-// Licensed under the Apache License, Version 2.0 (see LICENSE.md).
-package rsc.pretty
-
-import rsc.symtab._
-
-object PrettySymtab {
-  def str(p: Printer, x: Symtab): Unit = {
-    ???
-  }
-
-  def repl(p: Printer, x: Symtab): Unit = {
-    ???
-  }
-}
